컨테이너에 api 관련 행동 위임
